+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Release 0.95.0
 
 - Enhance collaborators permission logic.
 - Implemented threshold-based alerts for price estimates and quotas.
+- Prevent resource provisioning if total estimated cost of resource and project is over limit.
 
 Release 0.94.0
 --------------